conan integration of gRPC
It builds correctly for both clang and gcc based on the currently selected compiler. However, the resulting applications are huge when they are compiled in debug.

Having a way to build gRPC in release, no matter the application is in debug or not will be needed.

Also finding a way to have the shared/dynamic libs working would be also great.